+++ src/CW4Archipelago/Plugin.cs
@@ -6,12 +6,12 @@ using Il2CppInterop.Runtime.Injection;
 
 namespace CW4Archipelago;
 
-[BepInPlugin("com.droha.cw4archipelago", "CW4 Archipelago", "0.1.0")]
+[BepInPlugin("com.droha.cw4archipelago", "CW4 Archipelago", "0.1.1")]
 public class Plugin : BasePlugin
 {
     public override void Load()
     {
-        Log.LogInfo("CW4 Archipelago v0.1.0 loading");
+        Log.LogInfo("CW4 Archipelago v0.1.1 loading");
 
         var mcnet = typeof(Archipelago.MultiClient.Net.ArchipelagoSessionFactory).Assembly.GetName();
         Log.LogInfo($"Archipelago.MultiClient.Net {mcnet.Version} available");